refactor(oxfmt/lsp): move lsp code to `apps/oxfmt`, remove lsp formatter feature flag (#17462)
> This PR refactors the LSP formatter code by moving it from crates/oxc_language_server to apps/oxfmt, removing the formatter feature flag from the language server crate.

This is now possible, because `oxc_language_server` is not anymore shipped https://github.com/oxc-project/oxc/pull/17457
