+++ apps/oxfmt/src/lsp/tester.rs
@@ -2,11 +2,8 @@ use std::{fmt::Write, path::PathBuf};
 
 use tower_lsp_server::ls_types::{TextEdit, Uri};
 
-use crate::{
-    ToolRestartChanges,
-    formatter::server_formatter::{ServerFormatter, ServerFormatterBuilder},
-    tool::Tool,
-};
+use crate::lsp::server_formatter::{ServerFormatter, ServerFormatterBuilder};
+use oxc_language_server::{Tool, ToolRestartChanges};
 
 /// Given a file path relative to the crate root directory, return the absolute path of the file.
 pub fn get_file_path(relative_file_path: &str) -> PathBuf {
