+++ apps/oxfmt/test/fixtures/lsp/custom_config_path/semicolons-as-needed.ts
