+++ apps/oxfmt/test/fixtures/lsp/ignore-file/.prettierignore
