+++ apps/oxfmt/test/fixtures/lsp/ignore-file/ignored.ts
