+++ apps/oxfmt/test/fixtures/lsp/ignore-pattern/.oxfmtrc.json
