+++ apps/oxfmt/test/fixtures/lsp/ignore-pattern/ignored.ts
