+++ apps/oxfmt/test/fixtures/lsp/ignore-pattern/not-ignored.js
