+++ apps/oxfmt/test/fixtures/lsp/root_config/.oxfmtrc.json
