+++ apps/oxfmt/test/fixtures/lsp/root_config/semicolons-as-needed.ts
