+++ crates/oxc_language_server/Cargo.toml
@@ -28,15 +28,12 @@ doctest = false
 oxc_allocator = { workspace = true, optional = true }
 oxc_data_structures = { workspace = true, features = ["rope"], optional = true }
 oxc_diagnostics = { workspace = true, optional = true }
-oxc_formatter = { workspace = true, optional = true }
 oxc_linter = { workspace = true, optional = true }
-oxc_parser = { workspace = true, optional = true }
 
 #
 env_logger = { workspace = true, features = ["humantime"] }
 futures = { workspace = true }
 ignore = { workspace = true, features = ["simd-accel"], optional = true }
-json-strip-comments = { workspace = true }
 log = { workspace = true }
 papaya = { workspace = true }
 rustc-hash = { workspace = true }
@@ -50,20 +47,12 @@ insta = { workspace = true }
 tokio = { workspace = true, features = ["rt-multi-thread", "io-util", "macros"] }
 
 [features]
-default = ["linter", "formatter"]
+default = ["linter"]
 linter = [
   "dep:oxc_allocator",
   "dep:oxc_data_structures",
   "dep:oxc_diagnostics",
   "dep:oxc_linter",
-  #
-  "dep:ignore",
-]
-formatter = [
-  "dep:oxc_allocator",
-  "dep:oxc_data_structures",
-  "dep:oxc_formatter",
-  "dep:oxc_parser",
   #
   "dep:ignore",
 ]
