+++ crates/oxc_language_server/src/lib.rs
@@ -4,20 +4,16 @@ use tower_lsp_server::{LspService, Server, ls_types::ServerInfo};
 mod backend;
 mod capabilities;
 mod file_system;
-#[cfg(feature = "formatter")]
-mod formatter;
 #[cfg(feature = "linter")]
 mod linter;
 mod options;
 #[cfg(test)]
 mod tests;
 mod tool;
-mod utils;
+pub mod utils;
 mod worker;
 
-use crate::backend::Backend;
-#[cfg(feature = "formatter")]
-pub use crate::formatter::ServerFormatterBuilder;
+pub use crate::capabilities::Capabilities;
 #[cfg(feature = "linter")]
 pub use crate::linter::ServerLinterBuilder;
 pub use crate::tool::{Tool, ToolBuilder, ToolRestartChanges};
@@ -36,7 +32,11 @@ pub async fn run_server(
     let stdout = tokio::io::stdout();
 
     let (service, socket) = LspService::build(|client| {
-        Backend::new(client, ServerInfo { name: server_name, version: Some(server_version) }, tools)
+        crate::backend::Backend::new(
+            client,
+            ServerInfo { name: server_name, version: Some(server_version) },
+            tools,
+        )
     })
     .finish();
